modify functions

+++ ft_strlcpy.c
@@ -3,17 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tiacovel <tiacovel@student.42berlin.d      +#+  +:+       +#+        */
+/*   By: tiacovel <tiacovel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/28 14:18:52 by tiacovel          #+#    #+#             */
-/*   Updated: 2023/08/28 15:34:59 by tiacovel         ###   ########.fr       */
+/*   Updated: 2023/11/22 10:51:11 by tiacovel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcpy(char *dest, const char *src, unsigned int size)
+#include <stdlib.h>
+
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
@@ -38,8 +40,8 @@ int	main(void)
 	char	src[20] = "Hello";
 	char	dest42[20] = " world";
 	char	destSTD[20];
-	int size1;
-	int size2;
+	size_t size1;
+	size_t size2;
 
 	size1 = ft_strlcpy(dest42, src, 4);
 	size2 = strlcpy(destSTD, src, 4);
